requirements.txt
# Core dependencies
boto3==1.26.137
botocore==1.29.165
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.3
idna==3.10
jmespath==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
requests==2.31.0
s3transfer==0.6.2
six==1.16.0
urllib3==1.26.20

# GCP dependencies
google-cloud-speech>=2.21.0
google-cloud-storage>=2.10.0

# Azure dependencies (optional)
# azure-cognitiveservices-speech>=1.31.0

# Local Whisper dependencies (optional)
# faster-whisper>=0.10.0
# torch>=2.0.0

# Testing dependencies
pytest>=7.4.0
pytest-mock>=3.11.0
# Local Whisper dependencies
faster-whisper>=0.10.0
torch>=2.0.0